+++ package.json
@@ -26,7 +26,7 @@
     "test:functional:existing-journey-e2e": "NODE_TLS_REJECT_UNAUTHORIZED=1 codeceptjs run --config ./src/test/functional/codecept.conf.js --steps --grep @existing-journey-e2e",
     "test:functional:new-journey-errors": "NODE_TLS_REJECT_UNAUTHORIZED=1 codeceptjs run --config ./src/test/functional/codecept.conf.js --steps --grep @new-journey-errors",
     "test:crossbrowser": "./bin/run-crossbrowser-tests.sh",
-    "test-crossbrowser-e2e": "NODE_ENV=testing ALLOW_CONFIG_MUTATIONS=true LOG_LEVEL=OFF NODE_PATH=. yarn codeceptjs run-multiple ${BROWSER_GROUP:-'--all'} -c src/test/crossbrowser/saucelabs.conf.js",
+    "test-crossbrowser-e2e": "NODE_TLS_REJECT_UNAUTHORIZED=1 codeceptjs run-multiple ${BROWSER_GROUP:-'--all'} -c src/test/crossbrowser/saucelabs.conf.js",
     "cichecks": "yarn add puppeteer && yarn build && yarn rebuild puppeteer && yarn lint && yarn test && yarn test:routes",
     "prepare": "husky install",
     "pvmount:demo": "REMOTE_ENV=cui-demo NAMESPACE=cui node bin/mount-property-volume.js",
